[Silver I] Title: 징검다리 건너기 (large), Time: 220 ms, Memory: 14992 KB -BaekjoonHub

+++ 백준/Silver/22871. 징검다리 건너기 （large）/징검다리 건너기 （large）.java
@@ -1,7 +1,6 @@
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 import java.util.StringTokenizer;
 
 public class Main {
@@ -11,18 +10,18 @@ public class Main {
         int N = Integer.parseInt(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        long[] A = new long[N];
+        int[] A = new int[N];
         for (int i = 0; i < N; i++) {
-            A[i] = Long.parseLong(st.nextToken());
+            A[i] = Integer.parseInt(st.nextToken());
         }
 
         long[] dp = new long[N];
         dp[1] = 1 + Math.abs(A[1] - A[0]);
 
         for (int i = 2; i < N; i++) {
-            dp[i] = i * (1 + Math.abs(A[i] - A[0]));
+            dp[i] = i * (long)(1 + Math.abs(A[i] - A[0]));
             for (int j = 1; j < i; j++) {
-                dp[i] = Math.min(dp[i], Math.max(dp[j], (i-j) * (1 + Math.abs(A[i] - A[j]))));
+                dp[i] = Math.min(dp[i], Math.max(dp[j], (i-j) * (long)(1 + Math.abs(A[i] - A[j]))));
             }
         }
